Adding debugger and telescope, fix for lsp
The lsp config was overwritten by cmp. Merged lsp mappings
into cmp and disabled lsp config. When running lsp w/o cmp
this needs to be enabled again.
The debugger requires a debugger to be install. If a vscode
worksapce is configured it should work otherwise dbuggers
have to be installed seperately. Currently on the python
debugger dapter is install go and rust are not yet configured.